+++ NEWS.md
@@ -2,12 +2,12 @@
 ---------------------
 * General changes
     * Have `exprso` throw error if passed boolean outcomes
-    * [Make `ExprsPredict` and `RegrsPredict` sub-classes]
+    * Add new `RegrsPredict` class with two slots
 * Revise fs modules
     * Pass additional arguments to `fsPropd` function
     * Make `fsStats` run a bit faster
 * Revise build modules
-    * [Add `RegrsModel` support to `build.` constructor]
+    * Add `RegrsModel` support to `build.` constructor
 * Revise pl modules
     * Add default argument to `calcMonteCarlo`
     * Add default argument to `calcNested`